Complete Code refactor
